# internal/app/httpsrv/v1api/docs/swagger.yaml
basePath: /v1
definitions:
  domain.Ticket:
    properties:
      createdAt:
        type: string
      id:
        type: string
      key:
        type: string
      queue:
        type: string
      summary:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  fwerr.Code:
    enum:
    - INTERNAL_ERROR
    - AUTH_ERROR
    - FORBIDDEN
    - DB_ERROR
    - BAD_REQUEST
    - NOT_FOUND
    type: string
    x-enum-varnames:
    - CodeInternalError
    - CodeAuthError
    - CodeForbidden
    - CodeDBError
    - CodeBadRequest
    - CodeNotFound
  httpresp.ErrResponse:
    properties:
      code:
        $ref: '#/definitions/fwerr.Code'
      message:
        type: string
    type: object
  httpresp.SuccessResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  v1api.infoResponse:
    properties:
      commit_sha:
        type: string
      version:
        type: string
    type: object
  v1api.oauth2CallbackResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token:
        type: string
      not-before-policy:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      session_state:
        type: string
      token_type:
        type: string
    type: object
  v1api.ticketGetResponse:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/v1api.ticketGetResponseData'
    type: object
  v1api.ticketGetResponseData:
    properties:
      item:
        $ref: '#/definitions/domain.Ticket'
      total:
        type: integer
    type: object
  v1api.ticketListRequest:
    properties:
      filter:
        $ref: '#/definitions/v1api.ticketListRequestFilter'
      offset:
        type: integer
      size:
        type: integer
    type: object
  v1api.ticketListRequestFilter:
    properties:
      key:
        type: string
    type: object
  v1api.ticketListResponse:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/v1api.ticketListResponseData'
    type: object
  v1api.ticketListResponseData:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Ticket'
        type: array
      total:
        type: integer
    type: object
externalDocs:
  description: wiki.ddd.ru
  url: https://wiki.ddd.ru/
host: localhost:8080
info:
  contact:
    name: API Support
  description: This is a server
  title: alerts API
  version: v1.0
paths:
  /auth/oauth2/callback:
    get:
      consumes:
      - application/json
      description: Callback
      parameters:
      - description: authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1api.oauth2CallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
      summary: callback
      tags:
      - oauth2
  /auth/oauth2/login:
    get:
      description: Login
      parameters:
      - description: redirect url
        in: query
        name: redirect_url
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1api.oauth2CallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
      summary: login
      tags:
      - oauth2
  /auth/oauth2/logout:
    post:
      consumes:
      - application/json
      description: Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpresp.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
      security:
      - KeycloakAuth: []
      summary: logout
      tags:
      - oauth2
  /info:
    get:
      description: info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1api.infoResponse'
      summary: info
      tags:
      - info
  /ticket/{id}:
    get:
      consumes:
      - application/json
      description: User list
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1api.ticketGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
      summary: it's very to get all tickets
      tags:
      - tickets
  /ticket/list:
    post:
      consumes:
      - application/json
      description: User list
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1api.ticketListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1api.ticketListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.ErrResponse'
      summary: it's very to get all tickets
      tags:
      - tickets
securityDefinitions:
  KeycloakAuth:
    description: Keycloak authorization.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"